// src/api/orval/model/addOrUpdateOvpnFileConfigRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface AddOrUpdateOvpnFileConfigRequest {
  /**
   * @minimum 1
   * @maximum 2147483647
   */
  vpnServerId: number;
  /** @minLength 1 */
  vpnServerIp: string;
  vpnServerPort?: number;
  /** @nullable */
  configTemplate?: string | null;
}